implementation/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
structlog
redis
asyncpg
PyJWT
qdrant-client[grpc]
langchain
langchain-community
langchain-ollama
langgraph
sentence-transformers
httpx
python-jose[cryptography]
python-multipart
pdfplumber
python-docx
numpy
scipy
pandas
python-dotenv
prometheus-client
yfinance
